ws80211_utils: Fix HE and EHT PHY capabilities on Big Endian
The HE and EHT PHY Capabilities IEs are 11 and 9 bytes long,
respectively, so they're sent in the IE's original Little Endian
order, unlike the HT and VHT PHY capabilities, which are 2 and 4
bytes long and sent in the nl80211 message in native byte-order.

Luckily we only care about bits in the least significant byte, which
is sent first in LSB order, so we can just get the first byte and
not swap it.

Fixup 48de1716fd6501ced97482e25fcc8c03d51dd9bc

References:
        https://git.kernel.org/pub/scm/linux/kernel/git/jberg/iw.git/commit/?id=c41971e27a8359f88122593a2700f270f58cb2fa
        https://git.kernel.org/pub/scm/linux/kernel/git/jberg/iw.git/commit/?id=59660a349cf35903e951f99bdd8a74df063c912e