ws80211_utils: Make error handling more consistent
The various functions in ws80211_utils some times return -1
or NULL on failure, and sometimes return an error code. The
error codes are sometimes errnos, sometimes nl80211 errors
(on Linux), and sometimes ws80211 specific error numbers.

In the last case especially sometimes functions write directly
to standard error as well on unexpected libnl-3 errors, which is
no longer as bad as it was before #12222 was fixed, but still
should be avoided.

Add consistent return codes. Add a function to retrieve the error
message. For Linux libnl-3 error messages, store the last error
message in the nl_state struct and return it instead of writing
to stderr.