ws80211_utils: Return correct error from ws80211_iface_up
Instead of always returning -1, return -errno, since this error
code is eventually interpreted as a negative errno. Most of the time
the error message produced was correct, since the most likely place
to fail is trying to open a raw socket without CAP_NET_RAW, which
is a EPERM, which is typically 1. But returning the correct error
is preferable.