Clean up sync pipe read code.
Ensure that pipe_read_bytes() is never asked to read a number of bytes
that won't fit in an ssize_t, so that it will return either -1 on an
error or a positive number on success. (That should, in practice, always
be the case, as it'll either be reading a 4-byte message header or an
amount of data that is 2^24-1 bytes or less.)

First check for a return value of -1 before checking whether the return
value is equal to the number of bytes requested; that ensures that, when
doing said comparison, we can cast the number of bytes returned to
size_t and avoid a signed vs. unsigned comparison warning. (That's not
necessary when reading the 4-byte header, as the constant 4 is an int.)