ws80211_utils: Handle tuning 40 MHz in 6 GHz band
In the 6 GHz band, only HE (802.11ax/Wi-Fi 6) operation is allowed,
and the 40 MHz bands are non-overlapping. Despite this, some drivers
will report that they support both of HT40- and HT40+ on a given
frequency, even though that isn't true. Introduce a "40 MHz channel,
where center frequency must be specified" type, and use it
appropriately.

This eliminates almost all of the invalid tuning options from being
presented to the user in the 6 GHz band. If the driver reports that
a particular tune will definitely fail, it's better not to present
the option just to let it fail. (I still have an issue where my device
does not bother to report that due to regulatory restrictions in the
global operating class that 6 GHz channels 225, 229 and 233 can't be
used in 80 or 160 MHz node, and 233 not in 40 MHz mode either. But at
least we're doing the right thing here.)