diameter: Define custom entity in a valid fashion
The suggestion of what to do in Custom.xml did not produce valid XML.

This is an actually valid XML method of having a &Custom; entity
that itself includes other entities. The separate file should be
a parameter entity, which then defines a custom entity into which
the other entities are added.

This will maintain our practice of keeping the "dictionary.xml"
file as constant and allowing people to change only the "Custom"
file and add new files.

Using ".ent" is a more normal extension for this; might as well
change it as things are going to break anyway because the old
configuration wouldn't work.