log3gpp: Reduce use of static arrays.
Instead of statically allocating _three_ huge line buffers, combine them to a
single buffer in the dynamically allocated log3gpp_t private structure and a
much smaller buffer on the stack for probing the file type.  (The buffers
never needed to be separate; There is no risk of conflict by combining them in
this way.)  Similarly consolidate a few other static variables into the struct.

Reduces libwiretap's .bss by nearly 200 kBytes at the cost of 64 kBytes of heap
usage when the parser is active.  Also eliminates conflict for those static
buffers if we ever end up concurrently using multiple instances of this parser.