capture: Add a function to get the compiled BPF program for a capture filter
pcap_compile(3PCAP) notes:

"On Linux, if the pcap_t handle corresponds to a live packet capture, the
resulting filter program may use Linux BPF extensions." Due to privilege
separation, we likely only have the permissions to open a live handle
with pcap_create(3PCAP) or pcap_open_live(3PCAP) from dumpcap, for the
same reasons that the list of device and their capabilities can only be
obtained via dumpcap. It is useful to have a function to obtain the actual
filter that will be used on a live capture on a device.

Cf. https://ask.wireshark.org/question/37381/why-does-dumpcap-d-and-wireshark-capture-options-compile-bpfs-produce-different-output-for-the-same-capture-filter/