C12.22: Avoid UB if no base OID is given.
At initialization, the C1222 dissector reads a preference string to determine a
base OID, but the base OID remains as NULL (with length 0) if the preference
doesn't exist or is incorrectly formatted.  It then subsequently "memcpy"s that
base OID during packet dissection.  My system headers declare memcpy with the
nonnull attribute on its pointer arguments, so UBSAN complains -- even though
the size to copy is zero.

It's easy enough to simply skip the memcpy if no valid base OID was specified.