Qt: Get the BPF filter actually used in CompiledFilterOutput
On Linux, at least, (and now Windows with recent npcap) the compiled
capture filter for a live packet capture may use BPF extensions (e.g.,
when filtering for VLAN, because what the kernel hands to libpcap may
have the VLAN tag removed and placed elsewhere.) There are also some
capture filter expressions that can be valid for a live capture but
not when reading from a saved file (e.g., the Linux kernel can support
the ifindex, inbound, and outbound keywords on any live capture).
See pcap_compile(3PCAP) for details.

The only way to get the actual BPF filter used is to call dumpcap with
the actual interface name (and any necessary link-layer type), instead
of compiling a generic filter for the link-layer type. Note that while
this can fail with permission errors due to lack of privileges, it
already failed in the same scenario with bad permissions - the
difference is that the lack of permissions earlier led to an error
message about an illegal "-1" link-layer type for "Unknown". Now the
error message is more accurate.

We can't do this for pipes, standard input, or extcaps, which
aren't devices that dumpcap knows about.

For the capture syntax worker that is called constantly whenever the
text changes (as opposed to when a button is clicked), for now
mark capture filters that use such expressions as unknown (yellow/
deprecated). It's not clear if we want to hold open a bunch of pcap_t
handles (or open them repeatedly, which might spawn UAC prompts on
Windows.) That will at least allow filters with those keywords to be
compiled, and to be used in a capture.

The long term solution is probably something like holding dumpcap open
as a child process and sending messages back and forth, so it only has
to open devices once.