capture_sync: pass linktype when getting BPF filter
We need to pass in the linktype in order to pass it to dumpcap to
compile the BPF filter correctly.

Also, handle the SP_BAD_FILTER message type that can be returned from
dumpcap when attempting to compile a filter. (E.g., if compiling a
filter in DOCSIS, which doesn't support filtering.)

Make the primary and secondary message text pointers const, because
they're pointing directly into the received message and don't need
to be freed.