POSIX compatibility: Include lowercase Windows header on MinGW
Windows is "case-aware" in that even though it's case-insensitive by
default, NTFS file systems (and others) store case-sensitivity and it
is possible to force Windows into a case sensitive mode:
https://learn.microsoft.com/en-us/windows/wsl/case-sensitivity

The Windows header "BaseTsd.h" has that capitalization, officially, in
Windows, but is always lowercase in MinGW. If cross-compiling with MinGW
from a Linux system, or in some other situations, the include file can
be treated case-sensitively and fail.

One option is just to always include as lowercase, though that can break
in the rare case where someone has made Windows case-sensitive.