CMake: Use cmake_dependent_option for strato[shark] extcaps
cmake_dependent_option has the slightly different (and better)
behavior of automatically disabling the dependent option if the
necessary option is disabled, but keeping the old value internally
and restoring it if the necessary option is re-enabled.

So here, if stratoshark or strato is being built, and sshdig is
enabled, disabling stratoshark and strato will cause sshdig to
not be built, but later reenabling one of them will cause sshdig
to be built again.

It also hides impossible options in the GUI if that is being used.
There are other dependent options that it could be used for, but
it's less important there, since those are mostly platform options
that shouldn't change for the same set of build files. (I.e., whether
a system is Linux or WIN32 or Apple.)

https://cmake.org/cmake/help/latest/module/CMakeDependentOption.html