Add DECT NR+ MAC layer
DECT NR+ is specified in ETSI TS 103 636.

A preference is available to configure the Physical Header Field
type (40 bits or 80 bits) if the trace does not follow the spec
defined for WTAP_ENCAP_DECT_NR.

A preference is available to configure DLC PDU data type. The
default Automatic configuration will use heuristics to detect IPv6.

The specification uses different field lengths based on the value
of µ (subcarrier scaling factor). This is handled here by checking
the expected length for a message or IE.

The use of CVG Protocol Data Units is not implemented.