bthci: Remove redundant "Dissect all vendor cmds as Android" pref
There's already a Decode As Next Proto (vendor_dissector_table)
for BT HCI commands. That can be set to always dissect vendor-
specific commands as Android instead of using the actual manufacturer.
The Android dissector was added to that table along with the table
that uses the manufacturer ID. Thus, there's no need for a separate
preference that does exactly the same thing as the payload table and
just makes the code complicated. It's easier to be able to select
any vendor through Decode As than to have preferences for various
vendors.

This preference was introduced in fa3313b0c53840e95311a7ce3f859e7
at the same time as the Android dissector and has not been in a
released version, so it shouldn't need to be marked as obsolete.