falcodump: Exit with success if there are no interfaces
The tests expect a successful return code even if there are on
interfaces (even if the "always_present" is false.) That's what
androiddump does.

Both should perhaps return a special exit code for no interfaces
(WS_EXIT_NO_INTERFACES), though the extcap tests harness would need
to be changed to allow that exit code when no interfaces are
allowed.

This keeps tests from failing, e.g. on the Windows packages.