BT HCI: Revert separating Commands and Events payload tables
Currently, the registered Bluetooth HCI vendor tables handle both
Commands and Events, using the value of pinfo->p2p_dir to determine
whether the tvbuff passed in is a Command packet or an Event packet, as:

"HCI Command packets can only be sent to the Bluetooth controller,
HCI Event packets can only be sent from the Bluetooth controller..."
[Bluetooth Core Specification Vol 4, Part A, 2 Protocol, et passim]

Note that in the case of certain Events, Command Complete and
Command Status, the vendor dissector is not (necessarily) called
unless the Command Opcode embedded later in the Event turns out to
be a Vendor-Specific opcode. In each case, the entire Event (or
Command) packet is handed to the vendor dissector, so that the vendor
dissector can see the Event Code and handle Vendor-Specific, Command
Complete, and Command Status events differently. Similarly, the
entire Command packet beginning with opcode is handed to the vendor
dissector.  This causes adding the event code (or command opcode) and
parameter length multiple times to the tree.

Not all capture files and link-layer encapsulations necessarily have
a way to indicate direction, and not all might do so correctly.
So we might want to have multiple dissectors for each vendor, such as
one that always dissect as an Event and one that always dissects as a
Command. Or we might want to override the value of pinfo->p2p_dir before
calling the dissector tables. The beginnings of the former was done
by commit 55a739a67128080dd3998bc3430208860408fa1c, but it didn't
change the vendor dissectors distributed with Wireshark (and it
skipped the Event Code and parameter length when handing events
to the vendor Event dissector, which doesn't work to distinguish
the three event codes it can receive.)

Revert that commit change unless and until the vendor dissectors shipped
with are changed to match it. Right now they don't work with the new
table and handling.

Add comments because this is confusing.


(cherry picked from commit 78cf333c011a97e0cbf1a446680a41e3c580ebdb)

Co-authored-by: John Thacker <johnthacker@gmail.com>