capture_opts: Allow unoptimized capture filters on command line
Accept a long option "--no-optimize" (the same as tcpdump's long option)
not to optimize capture filters for all or any individual interface.
Implement the ability in dumpcap. Support the long option in dumpcap
and tshark.

It is not yet indicated in the GUI, so do not yet support it in Wireshark
(on the command line either.) It is also not yet supported for extcaps
(though adding it at least for sshdump should be possible, if sshdump is
configured to call dumpcap on a server with an earlier version of
dumpcap installed that does not have the option, then it will fail.)

Document it where it is supported. (Note that strato does not support
capture options, so capture options are currently commented out in the
man page, but capture options are processed on the command line to the
extent that an appropriate error message about the lack of capture
support can be given. Support for strato and stratoshark also depends
on the ability to use the option with extcaps.)

This should be an uncommonly used option, so do not use the short
option "-O" for this, which in any case is available for dumpcap
but not for tshark.

Ping #18190