Qt: Show unoptimized capture filters in CompiledFilterOutput
If the device is set to no-optimize (only possible from the command
line currently) then show the compiled capture filter as unoptimized.
Use the interface_t to describe the device. (It might be nice to have
a C++ wrapper around this at some point.)

Eventually we'd like to get the filters for multiple devices at once,
which will require a machine-readable output for the filter results
similar to that for the capabilities.

Ping #18190