capture: Handle empty filters correctly
Don't add an argument for an empty filter.
The check for an empty (non-NULL) string is not really necessary
if -f is the last argument, but would be necessary if performing
a check for more than one interface at a time.

This change is not needed before 19d655766de45944bb20abeabf19bc6863006709
because before that there's a pass to a QString and then back to a
const char* that causes the string to never be null, only empty.