HTTP: Don't save a pinfo->pool scoped map in file scoped data
If for some reason we have a supposed HTTP message with bogus data,
in that a field-line proceeds any start-line (thus violating RFC 9112),
do not store the pinfo->pool allocated header field map in the the
file scope allocated http_req_res_t for retrieval by later frames.
Instead, allocate a new map. (This loses all the fields that occurred
before the start-line, but as that's bogus anyway, that shouldn't be
a problem.)

Fix #20702. Should also fix #20701.

Add some comments. It might make sense in a subsequent change to only
allocate the file scoped header map after detecting an Upgrade header.
That would mean copying the map for any existing headers in the one
frame, but might be more straightforward logic, and would avoid
increasing persistent memory usage for HTTP requests and responses
that are not Upgrades and do not need to save the field data.