Graphs: Support floating-point scale factors
Add UAT support for storing and interpreting double-precision
floating point. Use that for the "Y Axis Scale Factor" in
graphs. This is forwards-compatible, but if a non-integer
number is stored in a UAT file and an older version of Wireshark
opens it, it will refuse to parse and fall back to the default
plots.

Fix #20667