LDA Neo trailer: make the names clearer.
LDA's website appear to contain zero instances of the string "ldaneo" in
Google's case-insensitive search.  Furthermore, the name doesn't
indicate that this is for an Ethernet trailer, unlike what's done for
at least some other trailers.

Rename the dissector source file, the protocol names, and
routine/variable/define names to reflect that.