Revert "LDA Neo trailer: make the names clearer."
This reverts commit a5b14603df5730d26c1227ca8f095e33b67abb91.

Try to do a rename and name fix in a way that preserves the history of
the file; Git may not be clueful enough to handle a change that renames
a file *and* modifies it (renames aren't entirely a thing in Gitland,
but newer versions seem to detect certain delete/add combinations as
really renames).

Start by undoing the change.