include: add the c128 128-bit arithmetic logic unit^Winclude library.
From Michael Clark's https://github.com/michaeljclark/c128; it provides
a "128-bit integer type with logic, shifts, arithmetic and bitmanip."

It is not yet used; the initial use will be for some places that have
128-bit time stamps; it could also be used for FT_{INT,UINT}128 types,
for example.

It's a header-only library, with #defines, type definitions, and static
inline routines.

This is the header files from that project with no changes other than
adding the copyright notice from integer.h to the other header files,
adding an SPDX entry after the copyright notice, and removing some
end-of-line whitespace, as the Wireshark builders require that.  This
establishes a baseline for any changes necessary to fit into Wireshark.

The LICENSE file for c128 claims, at the top, that it's the MIT license,
but the text in that file, and in integer.h, is closer to the ISC
license - https://opensource.org/license/isc-license-txt - than to the
MIT license - https://opensource.org/license/MIT. The *only* difference
between the license text here and the ISC license is that the first
sentence in the ISC license says "and/or" rather than "and". Thus, the
SPDX code here is ISC, not MIT.