Store an explicit file start time in the wtap header
Some file types, like BLF and LOG3GPP, have an explicit capture start
time in their header which is distinct from "the time stamp of the first
packet or other record with a time stamp." Store that in the wtap struct
similarly to how per-file encapsulation and timestamp precision is
stored. Use it for those two file formats.

For each record, the relative time to the capture start exists if and
only if the record has a time stamp, and the a capture start time
exists. Rather than having each record store the relative time to the
capture start (and whether it's valid), calculate it from the capture
start time for each record that has a time stamp similar to how relative
times to the first record or time reference frame are calculated.

We could possibly get rid of the value in the packet_info struct as well
and have the Plot dialog use the frame data function too.

Ping #20691, #20699