ClassicSTUN: Add padding heuristic
RFC 3489 aligns the attributes it defines on 32-bit boundaries
by padding them internally, such that att_length MUST always
be a multiple of 4. As RFC 5389 notes, this mechanism "never
worked for the few attributes that weren't aligned naturally
on 32 bit boundaries,"[1] such as USERNAME and PASSWORD. Note the
ClassicSTUN dissector dissects attribute types which were added
(and some later withdrawn) in draft-ietf-behave-rfc3489bis-00;
those also MUST be aligned.

There exist widely used classic STUN implementations that fail
to pad attributes; some (but not all?) of those use the RFC 5389
padding (introduced in draft-ietf-behave-rfc3489bis-04) whereby
att_length has the attribute length prior to padding and padding
up to 32 bit boundaries is added outside the attribute, despite
not using the Magic Cookie introduced in rfc3489bis-03 that would
cause the (RFC 5389 and later) STUN dissector to handle the frame.

Introduce a padding heuristic for the case where the attribute length
is not a multiple of 4. If there is only enough room in the message
length for padding but not another attribute, assume RFC 5389-style
padding. If there is room in the remaining message length for an
attribute either with or without padding, but the interpreting the
bytes with padding yields a known message type but interpreting the
bytes without padding does not, also assume RFC 5389-style padding.
Add a hf using the same style as used in the RFC 5389 STUN dissector.

1 - https://datatracker.ietf.org/doc/html/rfc5389#section-19


(cherry picked from commit dab2834a2bec1f76728866f134e20473ff3f4696)

Co-authored-by: John Thacker <johnthacker@gmail.com>