H.248 Annex E: Fix ToneDetection Tone ID List dissection
This is a sublist of Enumeration, which maps to a SEQUENCE OF
INTEGERS in ASN.1, though the main H.248 dissector takes care of
the SEQUENCE OF part. The other Tone ID lists, like hf_h248_pkg_cg_sig_pt_tl,
are already h249_param_ber_integer. Note this fixes the last parameter,
because implicit is a boolean; to use h248_param_uint_item the last
parameter be a length.

This caused a problem once gbooleans (which are actually 32-bit
integers) were replaced with real stdbool bools.

Fix #20666.