MQ: Eliminate the possiblity of overflow in MQINQ/MQSET
MQ officially designates certain values as MQLONG (which is a signed
32-bit integer) even though negative values are impossible. The
fields are already FT_UINT32; retrieve the values that way with
proto_tree_add_item_ret_uint. Test the sizes to make sure that
they won't overflow an int offset, though, similar to know but
without undefined behavior.

MQINT allows integer and character selectors to be interleaved; only
the relative order of the same type is important. Save the integer
selectors in order into a wmem_array as they are encountered to
eliminate the backtracking and extra bookkeeping for the offset.

Tested with the file in #9603.

Fix #20715