wscbor: Avoid UB in wscbor_chunk_read
When reading a chunk, truncate the length not to INT_MAX,
but to the value that would keep the offset to INT_MAX.
That will truncate and not throw an exception on the
current chunk, but will throw an exception when reading
the head for the next chunk.

Free the newly created chunk that won't be returned in
the case of exceptions.

Also, since wscbor_chunk_read can (and always could) throw
bounds errors, ignore those errors in bundle protocol when
determining whether or not to accept a frame.

Fix #20706