SIP: Fix line end offset when parsing Request-line
After advancing the the offset past the Method (and following SP),
the length advanced should be subtracted from the remaining line
length.

In practice there don't seem to be any errors from this currently,
as the end offset gets mostly used with tvb_find_uint8, which still
returns the correct values and doesn't throw exceptions if a too
long length is passed in.