format_units: Produce consistent results with the x87 FPU
GCC doesn't follow the C standard with regards to rounding 80-bit
intermediate values to the appropriate precision upon assignment and
casts, e.g. when using the x87 FPU on x86 processors without SSE2,
at least when compiling using -gnu11 or other -gnu dialects of C.
(It does by default with the dialects without extension, but then
we'd need to define _DEFAULT_SOURCE or similar in a number of places
where POSIX functions are used.)

This only matters for marginal values at certain boundaries, and
it's ultimately unimportant which value is used, but it makes the
test results consistent.

Fix #20720