Mongo: Avoid infinite loop in dissect_op_msg_section
If the size of a a OP_MSG data section is indicated as -1, that
leads to advancing the offset by section_len + 1, or zero, which
causes an infinite loop.

The total message and section lengths in Mongo are signed int32s;
it is impossible for them to be negative, and impossible for the
section length to be INT_MAX (since the message length includes
the length of the four byte headers and flag bits.)

Throw an error to avoid the offset moving backwards, an infinite loop,
or signed integer overflow.

Also update some URLs to their new locations.

Fix #20724.

(backported from commit 1ec4709cab382f7077ba66d2e382c2e75ce335c1)