SMB: Prettify the SMB Trans reassembly
Show a link from all fragments to the frame where the Transaction
payload is reassembled.
Cleanup the printout of other details as well.

Fragmented Transaction PDUs are exceptionally rare and you virtually
need to modify samba ot the client in order to trigger them
but  we should make the dissection a little more useful
nevertheless.
Though, as it is a very rare niche packet type for an obsolete
protocol, not invest too much time into it.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>