packet-darwin: added decoding of the drop reason code
Adds the ability to decode the "drop reason code" into a human readable
description, as well as to filter by the "drop component" and "drop
domain".

The mapping of the "drop reason" to the human readable description is
contained in the header file
"epan/dissectors/packet-darwin-droptap-msg.h", which roughly follows the
structure of the file "bsd/net/droptap.h" in the source code of the
"xnu" kernel. This mapping was donated by Apple Inc. under the MIT
license.


(cherry picked from commit db8bbcf92900435cafae3763c7c213e668191519)

Co-authored-by: Omer Shapira <omer_shapira@apple.com>