packet-cql.c: first part of supporting CQLv5 support
This protocol is a bit tricky - it starts with previous versions' negotiation, then moves to the new envelope format,
only to have the older frame format within...
To support it, we keep the protocol version and where the negotiation completed in the conversation.

This initial commit supports (tested) just the LZ4 compressed format of it.

Fixes: https://gitlab.com/wireshark/wireshark/-/issues/20646
Signed-off-by: Yaniv Kaul <mykaul@gmail.com>