EAX: Move a static struct onto the stack.
There's no need to use static storage for the decryption working space; No
state need be maintained across calls to Eax_Decrypt.  Moving it to the stack
frees up a tiny amount of .bss but most importantly eliminates the risk of
concurrency issues.

Tested using the capture attached to issue #5531.