XML: Use libxml2 function instead of struct elements directly
xmlValidPotentialChildren gets the list of potential children exactly
as a we want, without having to access internal struct members. (We
have to impose a maximum limit, but that's ok and it has the virtue
of limiting the recursion.)

Both this function and accessing the struct members were deprecated
in libxml2-2.15.0, but this function has been undeprecated for the
next release.


(cherry picked from commit 648b1603f6d7c6a21e4fde11dc6a82febcf79c14)

Co-authored-by: John Thacker <johnthacker@gmail.com>