IANA-IP: Fix missing braces.
The braces in the data structures in epan/iana-ip-data.c are ... numerous.  And
there's a section where we switch off the "missing-braces" diagnostic with a
comment suggesting that GCC has a problem here.

GCC appears to be correct; We were actually missing some braces, but Clang
accepted it without complaint.

Add the braces and remove the deactivation of the compiler diagnostic.  Both
GCC and Clang seem to accept the result.

To explain: The ipv6 field is a struct of type ipv6_addr_and_prefix; the
initializer for that struct is already demarcated by braces.  The first field
of that struct is of type ws_in6_addr, also a struct; a set of braces already
demarcates that initializer too.  The sole element of that struct is an array
of 16 bytes; initialization of that array requires braces, but those were
missing.