file: rename read_record to reflect that it doesn't read a record.
It doesn't read a record, it adds an *newly-read* record to the record
list if the record passes the read filter or if there is no read filter.

Also, as nobody who calls it cares any more whether the record was added
to the list (as of f210edeaf04c17da1e209518aae0e009e1164a73), make it
return void rather than bool.