test: Serialize Lua numbers with the C numeric locale
While Lua tonumber can accept the current locale's decimal point in
addition to period, load()ing an entire string uses the Lua lexer
and only accepts a period. Our test serializes an entire table, so
using, e.g., a comma as in the German locale makes the lexer unable
to parse a table correctly when unmarshaling.

If serializing, temporarily set the numeric locale to the C locale.

Add a test that explicitly sets the locale to German to the list
of lua tests in different locales.

Fix #20749