JSON-3GPP: Fix a few GRegex-related memory leaks.
Release the GMatchInfo using g_match_info_free, and g_free() the matched
substrings too.  With the GMatchInfo object now releasing its own reference to
the GRegex object, we must no longer "unref" the GRegex: we still hold a
reference to it in a static variable for later re-use.  (It was only the leak
of the reference held by the GMatchInfo object that was preventing further
usage of the GRegex from becoming a use-after-free.)