pidl: NULL is not a valid value for TYPE alignment in cnf files
A TYPE directive in a conformance file has to include an alignment
value, but it must be numeric. "NULL" isn't a valid value; with Perl
warnings on, it will still be treated as 0 but issues a warning:

Argument "NULL" isn't numeric in numeric lt (<) at /home/johnthacker/wireshark/tools/pidl/lib/Parse/Pidl/NDR.pm line 477.

Replace NULL with 0 where it appears in the cnf files. This doesn't
change the generated dissectors, but quiets warnings.